Update mixins to use brackets
This updates the with statment to require brackets. As part of this
I had to rethink how shapes worked. Previously we had fairly rigid
detection requirements for shapes that did / didn't have bodies and
what was allowed in those bodies. Since you can't have two subsequent
multiline rules, this presented some extreme difficulties. Instead,
the declaration of the shape and the shape body are now split. In
practice this means you theoretically have disconnected shape bodies.
This is somewhat unfortunate, but ultimately it's fine. This is
a syntax highlighter, not a parser. Perfectly enforcing the syntax
is a secondary goal to perfectly highlighting it. This isn't uncommon
either, you can for instance write "def def def" in a python file and
it will be highlighed by vs code just fine even though it won't parse
properly.