Fix tests relying on .smithy.lsp.log
Most of the test suites read the old .smithy.lsp.log file that the
language server could be configured to write out. Since that was
removed, we need a different way to do test assertions. I tried to
see if we could look at the trace of messages between server <->
vscode client, but wasn't able to find a way. So I simplified the
assertions for these test suites, which should be ok because we
probably don't want to rely on logs anyways for testing, and more
robust testing is done on the server side (our extension doesn't
have that much functionality to test).

I also had to add `sources` to each of the smithy-build.json files
used for tests, so the language server could properly load those
test projects.