Make couriser download jdk 21
By default, couriser will use AdoptOpenJDK 8 to run the server, so we
needed to tell it to use jdk 21. Coursier will download the JDK if
necessary. It uses an index to figure out where to download it from:
https://get-coursier.io/docs/2.0.6/cli-java#jvm-index. The version of
coursier we are using in the extension has an old version of the index,
so we needed to provide an updated index for coursier to pull from.
Also, we needed to tell coursier specifically which jdk to use (not just
21), because by default it tries to use AdoptOpenJDK, when there's no
AdoptOpenJDK 21 (it is under the name Adoptium) now.

This is a temporary solution, meant to avoid downloading a second version
of coursier which had an updated index and could properly find the right
jdk to download. In the future, we will ship the language server as a
standalone executable, and can remove coursier altogether.

I also updated the tests to stop using timeouts, so we don't have to worry
about an inconsistent download time in CI.