Add run command
Adds a config option to specify a specific command to run the language
server. When not set, the extension falls back to running with coursier.
By default, the config option is not set.

I need to clean up the config option a bit, since when you specify a
specific command, you can just use stdio, meaning the arguments passed
to the language server should just be "0". So really no args should be
set in the config option.