Cleanup client/server options
Refactored language client options and server config sent on
initialization, and removed unnecessary config.

We had some logic to figure out the root directory and workspace folder
which was used to refine which files the extension should care about,
but I think this is now unnecessary because the language server will
just find project roots anyway. There was also an extension config
option to set the root path, which should be unnecessary for the same
reason, so I removed it. The client still sends the workspace folder and
root path in the initialize request, so we aren't losing any information
here.

I also removed the file system watcher, as the server will dynamically
register that.