Cleanup coursier setup
Two things:

Some refactoring to use execFile instead of spawn for checking if
coursier is available, which is more lightweight.

Changed the way the server is run with coursier to have vscode run the
coursier command directly. Previously, we created a local tcp server
that acted as a proxy for the language server, which we ran in a
separate process. But I don't think this extra layer of indirection is
necessary, since vscode's lsp api can be given a command and some args
to use to start the server.