Cleanup config options
Changed the prefix of config properties from `smithyLsp` to just
`smithy` (or `smithy.server`, where applicable). I kept the existing
config properties, but added a deprecation message to them, pointing to
the new property. We will keep these in the next release, but remove
them in the following release.

Small note about the `smithy.server.version` property: I didn't give it
a default value, since that would have been chosen over a user-defined
version in the old property. When we remove the old properties, we can
add the default to the new one.

I also added a simple api for getting config settings in the extension,
that prefers the new properties over the old.