Fix config defaults
For some reason, when a string config option isn't specified it gets a
default value of `""`, so when we checked the new config option, the
empty string value would get chosen instead. So I added a default of
null for the old options, and I also switched the existing defaults to
the new options, choosing the old options first if they're present.