Cleanup extension (#106)
* Add run command

Adds a config option to specify a specific command to run the language
server. When not set, the extension falls back to running with coursier.
By default, the config option is not set.

I need to clean up the config option a bit, since when you specify a
specific command, you can just use stdio, meaning the arguments passed
to the language server should just be "0". So really no args should be
set in the config option.

* Remove unecessary format provider

The language server tells vscode it provides a formatter through its
capabilities, so we don't need any client-side code.

* Refactor lsp extensions

Cleaned up the implementations and organization of the
smithy/jarFileContents and smithy/selectorCommand lsp extension
handlers.

* Cleanup client/server options

Refactored language client options and server config sent on
initialization, and removed unnecessary config.

We had some logic to figure out the root directory and workspace folder
which was used to refine which files the extension should care about,
but I think this is now unnecessary because the language server will
just find project roots anyway. There was also an extension config
option to set the root path, which should be unnecessary for the same
reason, so I removed it. The client still sends the workspace folder and
root path in the initialize request, so we aren't losing any information
here.

I also removed the file system watcher, as the server will dynamically
register that.

* Cleanup coursier setup

Two things:

Some refactoring to use execFile instead of spawn for checking if
coursier is available, which is more lightweight.

Changed the way the server is run with coursier to have vscode run the
coursier command directly. Previously, we created a local tcp server
that acted as a proxy for the language server, which we ran in a
separate process. But I don't think this extra layer of indirection is
necessary, since vscode's lsp api can be given a command and some args
to use to start the server.

* Cleanup config options

Changed the prefix of config properties from `smithyLsp` to just
`smithy` (or `smithy.server`, where applicable). I kept the existing
config properties, but added a deprecation message to them, pointing to
the new property. We will keep these in the next release, but remove
them in the following release.

Small note about the `smithy.server.version` property: I didn't give it
a default value, since that would have been chosen over a user-defined
version in the old property. When we remove the old properties, we can
add the default to the new one.

I also added a simple api for getting config settings in the extension,
that prefers the new properties over the old.

* Fix config defaults

For some reason, when a string config option isn't specified it gets a
default value of `""`, so when we checked the new config option, the
empty string value would get chosen instead. So I added a default of
null for the old options, and I also switched the existing defaults to
the new options, choosing the old options first if they're present.